+++ client/src/components/Quiz.js
@@ -1,18 +1,10 @@
 import React  , { useState, useEffect } from 'react';
 import './Quiz.css';
+import QuizForm from './QuizForm';
 
 
 
 const Quiz = ({wordList}) => {
-
-    const[questionOneAnswer , setQuestionOneAnswer] = useState(null);
-    const[questionTwoAnswer , setQuestionTwoAnswer] = useState(null);
-    const[questionThreeAnswer , setQuestionThreeAnswer] = useState(null);
-    const[questionFourAnswer , setQuestionFourAnswer] = useState(null);
-    const[questionFiveAnswer , setQuestionFiveAnswer] = useState(null);
-
-    const[formSubmit, setFormSubmit] = useState()
-    
 
     const copyWordList = [...wordList]
 
@@ -67,146 +59,17 @@ const Quiz = ({wordList}) => {
     questionFiveWrongAnswersList.splice((questionFiveWrongAnswersList.indexOf(q5WrongA)), 1);
     const q5WrongB = questionFiveWrongAnswersList[Math.floor(Math.random()*questionFiveWrongAnswersList.length)];
 
-const handleChangeQ1 = (event) => {
-    setQuestionOneAnswer(event.target.value)
-}
-const handleChangeQ2 = (event) => {
-    setQuestionTwoAnswer(event.target.value)
-}
-const handleChangeQ3 = (event) => {
-    setQuestionThreeAnswer(event.target.value)
-}
-const handleChangeQ4 = (event) => {
-    setQuestionFourAnswer(event.target.value)
-}
-const handleChangeQ5 = (event) => {
-    setQuestionFiveAnswer(event.target.value)
-}
 
-const handleSubmit = (event) => {
-    event.preventDefault();
-    console.log(event);
-}
 
     return (
-        <div id="quiz">
-            <h3 id="topic-heading">Quiz</h3>
-            <p>Have fun with this little quiz to test your Spanish!</p>
-            <form onSubmit={handleSubmit}>
 
-                {/* QUESTION 1 */}
-                <div className="question-div q1-div">
-                    <img src={`${process.env.PUBLIC_URL}${questionOneWord.img}`} alt="Quiz topic"/>
-                    <div className="question-text">
-                        <p className="question-number">Question 1</p>
-                        <p className="question">What is the Spanish word for {questionOneWord.english}?</p>
-
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-1">{q1WrongA.translation}</label>
-                        <input type="radio" id="option-1" name="question-1" value="wrong" onChange={handleChangeQ1}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-2">{q1WrongB.translation}</label>
-                        <input type="radio" id="option-2" name="question-1" value="wrong" onChange={handleChangeQ1}/>
-                        
-                        {/* CORRECT ANSWER */}
-                        <label htmlFor="option-3">{questionOneWord.translation}</label>
-                        <input type="radio" id="option-3" name="question-1" value={questionOneWord.translation} onChange={handleChangeQ1}/>
-                    </div>
-                </div>
-
-                {/* QUESTION 2 */}
-                <div className="question-div q2-div">
-                    <img src={`${process.env.PUBLIC_URL}${questionTwoWord.img}`} alt="Quiz topic"/>
-                    <div className="question-text">
-                        <p className="question-number">Question 2</p>
-                        <p className="question">What is the English word for {questionTwoWord.translation}?</p>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-1">{q2WrongA.english}</label>
-                        <input type="radio" id="option-1" name="question-2" value="wrong" onChange={handleChangeQ2}/>
-                        
-                        {/* CORRECT ANSWER */}
-                        <label htmlFor="option-2">{questionTwoWord.english}</label>
-                        <input type="radio" id="option-2" name="question-2" value={questionTwoWord.english} onChange={handleChangeQ2}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-3">{q2WrongB.english}</label>
-                        <input type="radio" id="option-3" name="question-2" value="wrong" onChange={handleChangeQ2}/>
-                    </div>
-                </div>
-
-                {/* QUESTION 3 */}
-                <div className="question-div q3-div">
-                    <img src={`${process.env.PUBLIC_URL}${questionThreeWord.img}`} alt="Quiz topic"/>
-                    <div className="question-text">
-                        <p className="question-number">Question 3</p>
-                        <p className="question">What is the Spanish word for {questionThreeWord.english}?</p>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-1">{q3WrongA.translation}</label>
-                        <input type="radio" id="option-1" name="question-3" value="wrong" onChange={handleChangeQ3}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-2">{q3WrongB.translation}</label>
-                        <input type="radio" id="option-2" name="question-3" value="wrong" onChange={handleChangeQ3}/>
-                        
-                        {/* CORRECT ANSWER */}
-                        <label htmlFor="option-3">{questionThreeWord.translation} </label>
-                        <input type="radio" id="option-3" name="question-3" value={questionThreeWord.translation} onChange={handleChangeQ3}/>
-                    </div>
-                </div>
-
-                {/* QUESTION 4 */}
-                <div className="question-div q4-div">
-                    <img src={`${process.env.PUBLIC_URL}${questionFourWord.img}`} alt="Quiz topic"/>
-                    <div className="question-text">
-                        <p className="question-number">Question 4</p>
-                        <p className="question">What is the English word for {questionFourWord.translation}?</p>
-                        
-                        {/* CORRECT ANSWER */}
-                        <label htmlFor="option-1">{questionFourWord.english} </label>
-                        <input type="radio" id="option-1" name="question-4" value={questionFourWord.english} onChange={handleChangeQ4}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-2">{q4WrongA.english}</label>
-                        <input type="radio" id="option-2" name="question-4" value="wrong" onChange={handleChangeQ4}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-3">
-                            {q4WrongB.english} 
-                        </label>
-                        <input type="radio" id="option-3" name="question-4" value="wrong" onChange={handleChangeQ4}/>
-                    </div>
-                </div>
-
-                {/* QUESTION 5 */}
-                <div className="question-div q5-div">
-                    <img src={`${process.env.PUBLIC_URL}${questionFiveWord.img}`} alt="Quiz topic"/>
-                    <div className="question-text">
-                        <p className="question-number">Question 5</p>
-                        <p className="question">What is the Spanish word for {questionFiveWord.english}?</p>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-1">{q5WrongA.english}</label>
-                        <input type="radio" id="option-1" name="question-5" value="wrong" onChange={handleChangeQ5}/>
-                        
-                        {/* CORRECT ANSWER */}
-                        <label htmlFor="option-2">{questionFiveWord.translation} </label>
-                        <input type="radio" id="option-2" name="question-5" value={questionFiveWord.translation} onChange={handleChangeQ5}/>
-                        
-                        {/* WRONG ANSWER */}
-                        <label htmlFor="option-3">{q5WrongB.english}</label>
-                        <input type="radio" id="option-3" name="question-5" value="wrong" onChange={handleChangeQ5}/>
-                    </div>
-                </div>
-                <input type="submit" >
-
-                </input>
-
-                    
-                </form>
-            </div>
+        <QuizForm 
+        questionOneWord={questionOneWord} q1WrongA={q1WrongA} q1WrongB={q1WrongB}
+        questionTwoWord={questionTwoWord} q2WrongA={q2WrongA} q2WrongB={q2WrongB}
+        questionThreeWord={questionThreeWord} q3WrongA={q3WrongA} q3WrongB={q3WrongB}
+        questionFourWord={questionFourWord} q4WrongA={q4WrongA} q4WrongB={q4WrongB}
+        questionFiveWord={questionFiveWord} q5WrongA={q5WrongA} q5WrongB={q5WrongB}
+        />
         );
 }
